Separate build directories for linux and windows builds, some cleanup, better SetColorImage format/size validation

+++ src/gbd/rdram_file.c
@@ -47,7 +47,7 @@ rdram_file_pos(void)
 static bool
 rdram_file_addr_valid(uint32_t addr)
 {
-    return addr < rdram_file_size;
+    return addr < (unsigned long)rdram_file_size;
 }
 
 static size_t
